+++ lib/config.d.ts
@@ -1,0 +1,10 @@
+declare const _default: {
+    cssId: string;
+    panelClassName: string;
+    wrapperClassName: string;
+    defaultMenuWidth: number;
+    menuItemHeight: number;
+    menuItemDivideLineMargin: number;
+    defaultZIndex: number;
+};
+export default _default;